Switched to pytest parameterisation
