+++ tests/conftest.py
@@ -9,67 +9,9 @@ from pathlib import Path
 import git
 import pytest
 
-from pytest_flakefighters.database_management import (
-    TestException,
-    TestExecution,
-    TracebackEntry,
-)
-
 # pylint:disable=C0103
 pytest_plugins = "pytester"
 CURRENT_DIR = Path(__file__).parent
-
-
-@pytest.fixture(scope="function", name="test_execution")
-def _test_execution(flaky_reruns_repo):
-    return TestExecution(
-        outcome="failed",
-        exception=TestException(
-            name="AssertionError",
-            traceback=[
-                TracebackEntry(
-                    id=23,
-                    exception_id=1,
-                    path=os.path.join(
-                        flaky_reruns_repo.working_dir,
-                        "venv",
-                        "lib",
-                        "python3.11",
-                        "site-packages",
-                        "_pytest",
-                        "python.py",
-                    ),
-                    lineno=165,
-                    colno=13,
-                    statement="    result = testfunction(**testargs)",
-                    source="""
-@hookimpl(trylast=True)
-def pytest_pyfunc_call(pyfuncitem: Function) -> object | None:
-    testfunction = pyfuncitem.obj
-    if is_async_function(testfunction):
-        async_fail(pyfuncitem.nodeid)
-        funcargs = pyfuncitem.funcargs
-        testargs = {arg: funcargs[arg] for arg in pyfuncitem._fixtureinfo.argnames}
-        result = testfunction(**testargs)""",
-                ),
-                TracebackEntry(
-                    path=os.path.join(
-                        flaky_reruns_repo.working_dir,
-                        "flaky_reruns.py",
-                    ),
-                    lineno=21,
-                    colno=8,
-                    statement="        assert result",
-                    source="""
-        def test_create_or_delete(self):
-            flaky = FlakyReruns("test.txt")
-            result = flaky.create_or_delete()
-            assert result
-            """,
-                ),
-            ],
-        ),
-    )
 
 
 @pytest.fixture(scope="function", name="flaky_triangle_repo")
